Synthetic service feature flag (#119446)
