+++ x-pack/plugins/uptime/server/config.ts
@@ -11,7 +11,20 @@ import { schema, TypeOf } from '@kbn/config-schema';
 export const config: PluginConfigDescriptor = {
   schema: schema.maybe(
     schema.object({
-      index: schema.string(),
+      index: schema.maybe(schema.string()),
+      unsafe: schema.maybe(
+        schema.object({
+          service: schema.maybe(
+            schema.object({
+              enabled: schema.boolean(),
+              username: schema.string(),
+              password: schema.string(),
+              manifestUrl: schema.string(),
+              hosts: schema.arrayOf(schema.string()),
+            })
+          ),
+        })
+      ),
     })
   ),
 };
