+++ x-pack/plugins/uptime/server/plugin.ts
@@ -54,12 +54,14 @@ export class Plugin implements PluginType {
     });
 
     initServerWithKibana(
-      { router: core.http.createRouter() },
+      { router: core.http.createRouter(), config },
       plugins,
       ruleDataClient,
       this.logger
     );
-    core.savedObjects.registerType(umDynamicSettings);
+
+    registerUptimeSavedObjects(core.savedObjects, config);
+
     KibanaTelemetryAdapter.registerUsageCollector(
       plugins.usageCollection,
       () => this.savedObjectsClient
